Affordance_Modules/GLOVER-main/requirements.txt
bitsandbytes==0.41.1
einops==0.4.1
fastapi==0.100.1
gradio==3.39.0
markdown2==2.4.10
numpy==1.24.2
openai==0.27.8
opencv-python-headless==4.9.0.80
packaging==24.0
peft==0.4.0
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
requests==2.31.0
scipy==1.11.2
sentencepiece==0.2.0
shortuuid==1.0.11
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
triton==2.1.0
tensorboard==2.16.2
tqdm==4.64.1
transformers==4.31.0
uvicorn==0.23.2
deepspeed==0.14.2


